add sv_max_client_edicts
The server can be configured to use up to 2048 edicts, but clients will be disconnected if the server tries sending entity info for edicts above the default ~1600. Setting sv_max_client_edicts to a lower value will stop those entities being sent when a player joins. AddFullToPack is still called for every entity and a limit should be enforced there, too. The same goes for any decals, sounds, and other network messages which take entity indexes as a parameter.

Normally you would update the liblist.gam for the client side of your mod to fix this, but I want to run more edicts on a Half-Life server, not create a new mod.