add flags for skipping entity collision
Some flags now have multiple meanings, and so should only be used on entities where there will be no confusion on what the flag means (e.g. func_clip has no health so FL_IMMUNE_LAVA obviously doesn't apply).

Also, the oldorigin of an entity is set before executing a move, so that the game code can undo the move in a Touch function (required for func_pushable to prevent it getting stuck inside the player pushing it).